requirements.txt
transformers
torch
sentence-transformers
faiss-cpu
pyttsx3
simpleaudio
TTS
PyYAML
whisper
ffmpeg
openai-whisper
speechrecognition
sounddevice
scipy
numpy
faster-whisper